Mention ggplot2 and ggridges in sample code for completeness
If these aren't imported or otherwise available, code that uses the
package will error and complain that they aren't present.

This PR explicitly mentions these libraries in the examples.

An alternate approach would be to use the full namespace of
`ggplot2::autoplot` in the examples (as is done in some of the
Rmarkdown).