Use exact name of vignettes so instructions work
The original docs mentioned `vignette("Custom Simulations")` but this is
the title of the vignette, not the file name that can be used to launch
the vignette from the R console.

In order to work as a command in the R console, the name must be the
filename such as `vignette("custom")`.

Correct me or reject this PR if I'm wrong!